requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
